refactor(core): move the data store operation out of instructions
The view data store operation is not an instruction and shouldn't be
located in the instructions folder.
