+++ packages/core/src/render3/instructions/let_declaration.ts
@@ -12,8 +12,7 @@ import {TNodeType} from '../interfaces/node';
 import {HEADER_OFFSET} from '../interfaces/view';
 import {getContextLView, getLView, getSelectedIndex, getTView, setCurrentTNode} from '../state';
 import {getOrCreateTNode} from '../tnode_manipulation';
-import {load} from '../util/view_utils';
-import {store} from './storage';
+import {load, store} from '../util/view_utils';
 
 /** Object that indicates the value of a `@let` declaration that hasn't been initialized yet. */
 const UNINITIALIZED_LET = {};
