+++ packages/core/src/render3/instructions/storage.ts
@@ -9,17 +9,6 @@ import {HEADER_OFFSET, LView, TView} from '../interfaces/view';
 import {getContextLView} from '../state';
 import {load} from '../util/view_utils';
 
-/** Store a value in the `data` at a given `index`. */
-export function store<T>(tView: TView, lView: LView, index: number, value: T): void {
-  // We don't store any static data for local variables, so the first time
-  // we see the template, we should store as null to avoid a sparse array
-  if (index >= tView.data.length) {
-    tView.data[index] = null;
-    tView.blueprint[index] = null;
-  }
-  lView[index] = value;
-}
-
 /**
  * Retrieves a local reference from the current contextViewData.
  *
